// src/service/file/schemas/storageUsageResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

export interface StorageUsageResult {
  /** 사용자 계정의 id */
  accountId: number;
  /** 업로드한 자료 수 */
  count: number;
  /** 현재 사용한 용량(byte) */
  size: number;
  /** 계정에서 사용가능한 최대 용량(byte) */
  sizeAssigned: number;
}